chore: move TradeFlowContext to src/cow-react/modules/limitOrders/services/types.ts file
Add SafeBundleFlowContext
