+++ src/cow-react/modules/limitOrders/containers/LimitOrdersConfirmModal/index.tsx
@@ -4,7 +4,7 @@ import { Currency, CurrencyAmount } from '@uniswap/sdk-core'
 import { CloseIcon } from 'theme'
 import { CurrencyInfo } from '@cow/common/pure/CurrencyInputPanel/types'
 import { LimitOrdersConfirm } from '../../pure/LimitOrdersConfirm'
-import { TradeFlowContext } from '../../services/tradeFlow'
+import { TradeFlowContext } from '../../services/types'
 import TransactionConfirmationModal, { OperationType } from 'components/TransactionConfirmationModal'
 import { limitOrdersConfirmState } from '../LimitOrdersConfirmModal/state'
 import { useWalletInfo } from '@cow/modules/wallet'
