+++ src/cow-react/modules/limitOrders/containers/LimitOrdersWidget/limitOrdersPropsChecker.ts
@@ -2,7 +2,7 @@ import { CurrencyInfo } from '@cow/common/pure/CurrencyInputPanel/types'
 import { OnImportDismissCallback } from '@cow/modules/trade/hooks/useOnImportDismiss'
 import { RateInfoParams } from '@cow/common/pure/RateInfo'
 import { PriceImpact } from 'hooks/usePriceImpact'
-import { TradeFlowContext } from '@cow/modules/limitOrders/services/tradeFlow'
+import { TradeFlowContext } from '@cow/modules/limitOrders/services/types'
 import { areFractionsEqual } from '@cow/utils/areFractionsEqual'
 import { genericPropsChecker } from '@cow/utils/genericPropsChecker'
 import { getAddress } from '@cow/utils/getAddress'
