+++ src/cow-react/modules/limitOrders/containers/TradeButtons/index.tsx
@@ -2,7 +2,7 @@ import React, { useCallback } from 'react'
 import { Trans } from '@lingui/macro'
 import { useAtomValue } from 'jotai/utils'
 import { useSetAtom } from 'jotai'
-import { TradeFlowContext } from '@cow/modules/limitOrders/services/tradeFlow'
+import { TradeFlowContext } from '@cow/modules/limitOrders/services/types'
 import { limitOrdersSettingsAtom } from '@cow/modules/limitOrders/state/limitOrdersSettingsAtom'
 import { useLimitOrdersDerivedState } from '@cow/modules/limitOrders/hooks/useLimitOrdersDerivedState'
 import { useLimitOrdersFormState } from '../../hooks/useLimitOrdersFormState'
