+++ src/cow-react/modules/limitOrders/hooks/useHandleOrderPlacement.test.ts
@@ -1,7 +1,8 @@
 import { renderHook } from '@testing-library/react-hooks'
 import { PriceImpact } from 'hooks/usePriceImpact'
 import { useHandleOrderPlacement } from './useHandleOrderPlacement'
-import { tradeFlow, TradeFlowContext } from '@cow/modules/limitOrders/services/tradeFlow'
+import { tradeFlow } from '@cow/modules/limitOrders/services/tradeFlow'
+import { TradeFlowContext } from '@cow/modules/limitOrders/services/types'
 import { defaultLimitOrdersSettings } from '../state/limitOrdersSettingsAtom'
 import { limitOrdersRawStateAtom, updateLimitOrdersRawStateAtom } from '../state/limitOrdersRawStateAtom'
 import { useAtomValue, useUpdateAtom } from 'jotai/utils'
