+++ src/cow-react/modules/limitOrders/pure/LimitOrdersConfirm/index.cosmos.tsx
@@ -5,7 +5,7 @@ import { COW, GNO } from 'constants/tokens'
 import { SupportedChainId } from '@cowprotocol/cow-sdk'
 import { OrderKind } from '@cowprotocol/cow-sdk'
 import { OrderClass } from '@cowprotocol/cow-sdk'
-import { TradeFlowContext } from '../../services/tradeFlow'
+import { TradeFlowContext } from '../../services/types'
 import { LimitOrdersConfirm } from './index'
 import { LimitOrdersWarnings } from '@cow/modules/limitOrders/containers/LimitOrdersWarnings'
 import React from 'react'
