+++ src/cow-react/modules/limitOrders/pure/LimitOrdersConfirm/index.tsx
@@ -5,7 +5,7 @@ import { ButtonPrimary } from 'components/Button'
 import { CurrencyInfo } from '@cow/common/pure/CurrencyInputPanel/types'
 import { CurrencyPreview } from '@cow/common/pure/CurrencyInputPanel'
 import { LimitOrdersDetails } from '../LimitOrdersDetails'
-import { TradeFlowContext } from '../../services/tradeFlow'
+import { TradeFlowContext } from '../../services/types'
 import * as styledEl from './styled'
 import { LOW_RATE_THRESHOLD_PERCENT } from '@cow/modules/limitOrders/const/trade'
 import { PriceImpact } from 'hooks/usePriceImpact'
