+++ src/cow-react/modules/limitOrders/pure/LimitOrdersDetails/index.cosmos.tsx
@@ -3,7 +3,7 @@ import { COW, GNO } from 'constants/tokens'
 import { SupportedChainId } from '@cowprotocol/cow-sdk'
 import { OrderKind } from '@cowprotocol/cow-sdk'
 import { OrderClass } from '@cowprotocol/cow-sdk'
-import { TradeFlowContext } from '../../services/tradeFlow'
+import { TradeFlowContext } from '../../services/types'
 import { LimitOrdersDetails } from './index'
 import { defaultLimitOrdersSettings } from '@cow/modules/limitOrders/state/limitOrdersSettingsAtom'
 import { initLimitRateState } from '@cow/modules/limitOrders/state/limitRateAtom'
