+++ src/cow-react/modules/limitOrders/pure/LimitOrdersDetails/index.tsx
@@ -2,7 +2,7 @@ import React, { useMemo, useState } from 'react'
 import { InfoIcon } from 'components/InfoIcon'
 import * as styledEl from './styled'
 import styled from 'styled-components/macro'
-import { TradeFlowContext } from '@cow/modules/limitOrders/services/tradeFlow'
+import { TradeFlowContext } from '@cow/modules/limitOrders/services/types'
 import { isAddress, shortenAddress } from 'utils'
 import { RateInfoParams } from '@cow/common/pure/RateInfo'
 import { LimitOrdersSettingsState } from '@cow/modules/limitOrders/state/limitOrdersSettingsAtom'
